Fix windows pypy problem in tests
